// src/app/displays/listado_Documento_Pendiente_Convertir/interfaces/param-convert-doc.interface.ts
export interface ParamConvertDocInterface {
    pUserName:           string;
    pO_Documento:        number;
    pO_Tipo_Documento:   number;
    pO_Serie_Documento:  string;
    pO_Empresa:          number;
    pO_Estacion_Trabajo: number;
    pO_Fecha_Reg:        number;
    pD_Tipo_Documento:   number;
    pD_Serie_Documento:  string;
    pD_Empresa:          number;
    pD_Estacion_Trabajo: number;
}